writers: rename NetworkWriter -> TCPWriter
This commit renames NetworkWriter into TCPWriter, since that's all it
does now.

To make the change more obvious, this commit also leaves behind an
empty NetworkWriter class that just prints error messages telling
users to use TCPWriter and/or UDPWriter instead.
